Backend designer with execution functions.
The current upload is simply to exemplify the ability of the backend designer to have hexidecimally-based color selections. Note that execution functions are not safe, especially given they are used in a user-input based context. The rather raw design of the color canvas is not near its current state and is to be adjusted with time through a mixture of code automation and tkinter asset redesign.